Adds an ajaxNow extension method to support ajax calls on demand
The first use case of Blade was for event driven ajax.
The ajaxNow method supports immediate execution for use cases such as ajax calls on page load.
I had thoughts of making a "fake" event called now and just using ajaxOn, but that didn't feel right.
The downside of this approach is two wire up calls, one for "on" and one for "now".