Fixes issues with propagation by moving logic into executeAjax
Not sure how I feel about this, but it seems like the least amount of code duplication.
Still, I definitely think this could be done better.
The issue is that both ajaxNow and ajaxOn need to return based on propagation.
Funneling ajaxOn through ajaxNow loses the return value for ajaxOn.
Again, probably something obvious can be done and I am just not seeing it right now.